Rewrite the application-specific functions section. Fixes #1
This splits up the discussion about intermediaries and connectivity,
which were a rough fit to start with.

I extensively rewrote the section on intermediaries to more clearly
focus on protocol intermediaries, in the sense of draft-thomson-tmi,
though I don't cite that draft. I also changed the example to be about
encrypting protocol metadata, which I think is a more currently relevant
tradeoff than the HTTPS caching one (though I think ECH and QUIC are on
the right side of that tradeoff).